fix: update imports for renamed phase modules
Updated imports from old module names to new simplified names:
- phase_a_parameter_update -> phase_a
- phase_b_science_check -> phase_b
- phase_c_pydantic_report -> phase_c

Files updated:
- cli/wizard/validators/yaml_processor_integration.py
- data_model/validation/pipeline/phase_a_reporter.py
- data_model/validation/pipeline/phase_b_reporter.py
- data_model/validation/pipeline/phase_c_reporter.py
- test/validation_json_reporting/test_backward_compatibility.py
- test/data_model/test_data_model.py (comment only)
